Move initialUsers
